fastapi/requirements.txt
fastapi
requests
pydantic
python-dotenv
uvicorn
snowflake-connector-python